requirements.txt
roboticstoolbox-python==1.1.1
numpy==1.26.4
sympy==1.13.3
matplotlib==3.5.1
scipy
